+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Générateur SEDA : Ajout de la description dans les unités d'archivage (nécessite générateur SEDA v>0.3.0)
 - Ordre des onglets dans `pdf-generique` : L'onglet `signature` vient après l'onglet `parapheur` et non à la fin #1217
 - L'utilisation de deux transformations dans un type de dossier studio provoquait un warning #1262
+- Lors d'une transformation, le titre n'était pas mis à jour s'il était modifié par la transformation #1263
 
 # [3.1.0] - 2021-03-01